Fixed bug when determining number of columns/rows on blank sheets
Excel stores information in each worksheet about what part of the sheet is "used". We use this to determine the number of rows or columns in a spreadsheet. When the sheet is empty, though, Excel indicates the "dimension" as A1 (not as a full reference with a colon). So first, we need to account for this when calculating the 'used_area'. Excel also uses 'A1', though, when there is only a value in the cell A1. This commit checks for both of those scenarios and sets the number of rows/columns correctly depending on whether there is data or not.