(Bug) Fixing bug to deal with cells without an "r" attribute
Apparently the "r" attribute on "c" elements is optional and some MS products put out worksheets without the "r" element. There does not appear to be documentation on the standard way to deal with this situation, but we simply assume that you start with column A and move 1 column over within a row for each cell in that row.

Don't currently have a test added for this because I don't have a small enough sheet to put into the library.