(Bug) Did not iterate through all rows for worksheets without "dimensions"
We use the "dimension" tag in a worksheet to calculate the number of rows in that sheet. When it wasn't present, we simply assumed 1 cell of data was present. This was a bad assumption as not all programs that create these files set the dimension. The solution is to simply not rely on the number of rows. Iterating through the xml will stop when there are no more rows anyway, so no need to rely on the "used_area" attribute.

Note that the used_area attribute of worksheets is unreliable for certain spreadsheets.